requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
cryptography==39.0.2
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.52.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
-e git+https://github.com/forrest-herman/python-packages-common.git@e7e459dd3384af7b12b7bfb8ae98e9efb773d1b4#egg=google_calendar_integrations&subdirectory=google_calendar_integrations
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.3
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.13.0
httplib2==0.20.4
idna==3.3
lxml==4.9.1
msgpack==1.0.5
oauthlib==3.2.2
outcome==1.2.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.10
wsproto==1.1.0
